refactor(observability): rename observability.observation package to observability.tracing
