+++ embabel-agent-autoconfigure/embabel-agent-observability-autoconfigure/src/main/java/com/embabel/agent/autoconfigure/observability/ObservabilityAutoConfiguration.java
@@ -21,15 +21,15 @@ import com.embabel.agent.api.event.observation.LlmObservationContext;
 import com.embabel.agent.api.event.observation.ToolLoopObservationContext;
 import com.embabel.agent.observability.ObservabilityProperties;
 import com.embabel.agent.observability.mdc.MdcPropagationEventListener;
-import com.embabel.agent.observability.observation.ChatModelObservationFilter;
-import com.embabel.agent.observability.observation.EmbabelActionObservationConvention;
-import com.embabel.agent.observability.observation.EmbabelAgentObservationConvention;
-import com.embabel.agent.observability.observation.EmbabelLlmObservationConvention;
-import com.embabel.agent.observability.observation.EmbabelSpanEventListener;
-import com.embabel.agent.observability.observation.EmbabelToolLoopObservationConvention;
+import com.embabel.agent.observability.tracing.ChatModelObservationFilter;
+import com.embabel.agent.observability.tracing.EmbabelActionObservationConvention;
+import com.embabel.agent.observability.tracing.EmbabelAgentObservationConvention;
+import com.embabel.agent.observability.tracing.EmbabelLlmObservationConvention;
+import com.embabel.agent.observability.tracing.EmbabelSpanEventListener;
+import com.embabel.agent.observability.tracing.EmbabelToolLoopObservationConvention;
 import com.embabel.agent.api.event.observation.AgentInstrumentation;
 import com.embabel.agent.observability.metrics.EmbabelMetricsEventListener;
-import com.embabel.agent.observability.observation.MicrometerAgentInstrumentation;
+import com.embabel.agent.observability.tracing.MicrometerAgentInstrumentation;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.observation.ObservationRegistry;
 import org.slf4j.Logger;
